rename conftest to data
