+++ test/transformer/data.py
@@ -1,6 +1,8 @@
 import numpy as np
 import pandas as pd
-from sklearn.datasets import load_iris
+import pytest
+
+load_iris = pytest.importorskip(modname="sklearn.datasets.load_iris")
 
 
 iris = load_iris(as_frame=True)
