+++ test/transformer/pandas/test_FilterTF.py
@@ -1,7 +1,7 @@
 from dtoolkit.accessor.dataframe import cols as dataframe_cols  # noqa: F401
 from dtoolkit.accessor.series import cols as series_cols  # noqa: F401
 from dtoolkit.transformer import FilterTF
-from test.transformer.conftest import df_period
+from test.transformer.data import df_period
 
 
 def test_work():
