+++ test/transformer/pandas/test_GetTF.py
@@ -4,8 +4,8 @@ from pandas.testing import assert_frame_equal
 from pandas.testing import assert_series_equal
 
 from dtoolkit.transformer import GetTF
-from test.transformer.conftest import df_iris
-from test.transformer.conftest import feature_names
+from test.transformer.data import df_iris
+from test.transformer.data import feature_names
 
 
 @pytest.mark.parametrize("cols", [feature_names[0], feature_names])
