+++ test/transformer/pandas/test_ReplaceTF.py
@@ -1,5 +1,5 @@
 from dtoolkit.transformer import ReplaceTF
-from test.transformer.conftest import df_label
+from test.transformer.data import df_label
 
 
 def test_work():
